+++ api/blueprints/drbUtils.py
@@ -22,84 +22,88 @@ EXCLUDED_HEADERS = [
 
 
 @utils.route('/languages', methods=['GET'])
-def languageCounts():
-    esClient = ElasticClient(current_app.config['REDIS_CLIENT'])
+def get_languages():
+    response_type = 'languageCounts'
 
-    reqParams = APIUtils.normalizeQueryParams(request.args)
-    workCounts = reqParams.get('totals', ['true'])[0].lower() != 'false'
+    try:
+        es_client = ElasticClient(current_app.config['REDIS_CLIENT'])
 
-    langResult = esClient.languageQuery(workCounts)
+        request_params = APIUtils.normalizeQueryParams(request.args)
+        work_counts = request_params.get('totals', ['true'])[0].lower() != 'false'
 
-    languageList = APIUtils.formatLanguages(
-        langResult.aggregations, workCounts
-    )
+        language_query_results = es_client.languageQuery(work_counts)
 
-    logger.debug('Language list 200 OK on /utils/languages')
-
-    return APIUtils.formatResponseObject(200, 'languageCounts', languageList)
+        return APIUtils.formatResponseObject(
+            200, 
+            response_type,
+            APIUtils.formatLanguages(language_query_results.aggregations, work_counts)
+        )
+    except Exception as e:
+        logger.error(e)
+        return APIUtils.formatResponseObject(500, response_type, 'Unable to get language counts')
 
 
 @utils.route('/counts', methods=['GET'])
-def totalCounts():
-    dbClient = DBClient(current_app.config['DB_CLIENT'])
-    dbClient.createSession()
+def get_counts():
+    response_type = 'totalCounts'
 
-    totalResult = dbClient.fetchRowCounts()
+    try:
+        with DBClient(current_app.config['DB_CLIENT']) as db_client:
+            total_counts = db_client.fetchRowCounts()
 
-    totalsSummary = APIUtils.formatTotals(totalResult)
-
-    dbClient.closeSession()
-
-    return APIUtils.formatResponseObject(200, 'totalCounts', totalsSummary)
+        return APIUtils.formatResponseObject(200, response_type, APIUtils.formatTotals(total_counts))
+    except Exception as e:
+        logger.error(e)
+        return APIUtils.formatResponseObject(500, response_type, 'Unable to get total counts')
 
 
 @utils.route('/proxy', methods=['GET', 'POST', 'PUT', 'HEAD', 'OPTIONS'])
 @cross_origin(origins=os.environ.get('API_PROXY_CORS_ALLOWED', '*'))
-def getProxyResponse():
-    proxyUrl = request.args.get('proxy_url')
-    cleanUrl = unquote_plus(proxyUrl)
-    urlParts = urlparse(cleanUrl)
+def proxy_response():
+    try:
+        proxy_url = request.args.get('proxy_url')
+        clean_url = unquote_plus(proxy_url)
+        url_parts = urlparse(clean_url)
 
-    logger.info(f'Received {request.method} request to proxy url: {proxyUrl}')
+        logger.info(f'Proxying {request.method} request to {proxy_url}')
 
-    while True:
-        headResp = requests.head(
-            cleanUrl, headers={'User-agent': 'Mozilla/5.0'}
+        while True:
+            head_resp = requests.head(clean_url, headers={'User-agent': 'Mozilla/5.0'})
+
+            status_code = head_resp.status_code
+            
+            logger.info(f'HEAD response from {clean_url} returned status code {status_code}')
+
+            if status_code in [200, 204]:
+                break
+            elif status_code in [301, 302, 303, 307, 308]:
+                clean_url = head_resp.headers['Location']
+
+                if clean_url[0] == '/':
+                    clean_url = '{}://{}{}'.format(url_parts.scheme, url_parts.netloc, clean_url)
+            else:
+                logger.warn(f'Unable to proxy clean url {clean_url}')
+                clean_url = proxy_url
+                break
+
+
+        response = requests.request(
+            method=request.method,
+            url=clean_url,
+            headers={k: v for (k, v) in request.headers if k != 'Host'},
+            data=request.get_data(),
+            cookies=request.cookies,
+            allow_redirects=False
         )
 
-        statusCode = headResp.status_code
-        
-        logger.info(f'HEAD response from {cleanUrl} returned status code {statusCode}')
+        headers = [
+            (k, v) for (k, v) in response.headers.items()
+            if k.lower() not in EXCLUDED_HEADERS
+        ]
 
-        if statusCode in [200, 204]:
-            break
-        elif statusCode in [301, 302, 303, 307, 308]:
-            cleanUrl = headResp.headers['Location']
+        logger.info(f'Returning {request.method} response from {clean_url} with status code {response.status_code}')
 
-            if cleanUrl[0] == '/':
-                cleanUrl = '{}://{}{}'.format(
-                    urlParts.scheme, urlParts.netloc, cleanUrl
-                )
-        else:
-            logger.warn(f'Unable to proxy clean url {cleanUrl}')
-            cleanUrl = proxyUrl
-            break
-
-
-    resp = requests.request(
-        method=request.method,
-        url=cleanUrl,
-        headers={k: v for (k, v) in request.headers if k != 'Host'},
-        data=request.get_data(),
-        cookies=request.cookies,
-        allow_redirects=False
-    )
-
-    headers = [
-        (k, v) for (k, v) in resp.headers.items()
-        if k.lower() not in EXCLUDED_HEADERS
-    ]
-
-    logger.info(f'Returning {request.method} response from {cleanUrl} with status code {resp.status_code}')
-
-    return Response(resp.content, resp.status_code, headers)
+        return Response(response.content, response.status_code, headers)
+    except Exception as e:
+        logger.error(e)
+        return Response('Unable to proxy response', 500)