+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added local S3 docker container via localstack
 - Added error handling to citation API
 - Implemented Counter 5 reporting for view counts
+- Added error handling for utils API
 
 ## Fixed
 - Changed HATHI_DATAFILES outdated link in development, example, and local yaml files